Cleanup, enhances docs and refactors duplicate code. Fixes issues reported by SCA.
