+++ opennlp-grpc/opennlp-grpc-service/src/test/java/opennlp/OpenNLPServerIT.java
@@ -25,6 +25,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
@@ -97,8 +98,8 @@ public class OpenNLPServerIT {
 
   private static Path getConfig() throws URISyntaxException {
     return Paths.get(
-        Thread.currentThread().getContextClassLoader()
-            .getResource("config-test.ini")
+        Objects.requireNonNull(Thread.currentThread().getContextClassLoader()
+                .getResource("config-test.ini"))
             .toURI()
     ).toAbsolutePath();
   }
